nucleus-rapid-blog/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
appnope==0.1.2
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
cleanlab==1.0
cycler==0.11.0
debugpy==1.4.1
decorator==5.0.9
docutils==0.17.1
fonttools==4.28.3
frozenlist==1.2.0
idna==3.2
imagesize==1.3.0
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
multidict==5.2.0
nest-asyncio==1.5.4
numpy==1.21.1
packaging==21.3
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prettytable==2.2.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
scale-nucleus==0.3.2
scaleapi==2.2.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.2.1
Sphinx==4.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.0.0
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
yarl==1.7.2
zipp==3.5.0
